refactor(stellar-sdk): update schema service with encoder integration
- Import SorobanSchemaEncoder for structured schema support
- Update validators to use standardizedSchemas imports
- Add support for XDR and JSON schema formats
